add-wip: goodid flow in gooddapp
