+++ src/routes.tsx
@@ -10,7 +10,7 @@ const Swap = lazy(() => import('./pages/gd/Swap'))
 const Stakes = lazy(() => import('./pages/gd/Stake'))
 const Portfolio = lazy(() => import('./pages/gd/Portfolio'))
 const MicroBridge = lazy(() => import('./pages/gd/MicroBridge'))
-const Claim = lazy(() => import('./pages/gd/Claim'))
+const Claim = lazy(() => import('./pages/gd/Claim/'))
 const GoodId = lazy(() => import('./pages/gd/GoodId'))
 const BuyGd = lazy(() => import('./pages/gd/BuyGD'))
 const NewsFeedPage = lazy(() => import('./pages/gd/News'))
